Move `Repository` to its own file
